Stability improvements to a few tests.
